# gh-111139: Optimize math.gcd(int, int)