@vstinner 2024-01-10
@serhiy-storchaka: Would you mind to review this change?

> vstinner added the skip news label

I don't think that it's worth it to document such minor micro-optimization.